moved PartsSelling test into PartsCreation and Renamed this test
