+++ tests/acceptance/manager/PartsCest.php
@@ -3,14 +3,20 @@
 namespace hipanel\modules\stock\tests\acceptance\manager;
 
 use hipanel\helpers\Url;
+use Codeception\Example;
+use hipanel\tests\_support\Page\IndexPage;
+use hipanel\modules\stock\tests\_support\Page\part\SellModalWindow;
+use hipanel\tests\_support\Page\Widget\Input\Dropdown;
+use hipanel\tests\_support\Page\Widget\Input\Input;
 use hipanel\modules\stock\tests\_support\Page\part\Create;
 use hipanel\tests\_support\Step\Acceptance\Manager;
 
-class PartsCreationCest
+class PartsCest
 {
 
-    protected $createPage;
+    protected Create $createPage;
     protected $testOrderData;
+    protected array $sellData;
 
     public function _before(Manager $I)
     {
@@ -140,12 +146,59 @@ class PartsCreationCest
     }
 
     /**
+     * @dataProvider getSellData
+     */
+    public function ensureICanSellParts(Manager $I, Example $example): void
+    {
+        $partIndex      = new IndexPage($I);
+        $I->needPage(Url::to('@part'));
+        $this->sellData = iterator_to_array($example->getIterator());
+
+        $partIndex->filterBy(Input::asTableFilter($I, 'Serial'), 'MG_TEST_PART');
+        for ($i = 0; $i < count($this->sellData['prices']); $i++) {
+            $partIndex->selectTableRowByNumber($i + 1);
+        }
+        $I->click("//button[contains(text(), 'Sell parts')]");
+        $I->click("//a[text()='Sell parts']");
+        $I->waitForPageUpdate();
+
+        $sellModal      = new SellModalWindow($I);
+        $sellModal->fillSellWindowFields($this->sellData);
+        $I->pressButton('Sell');
+        $sellModal->seePartsWereSold();
+    }
+
+    public function ensureSellingBillWasCreated(Manager $I): void
+    {
+        $billPage = new IndexPage($I);
+
+        $I->needPage(Url::to('@bill'));
+
+        $this->filterTable($I);
+        $billPage->openRowMenuByNumber(1);
+        $billPage->chooseRowMenuOption('View');
+
+        $I->seeNumberOfElements('tr table  tr[data-key]', count($this->sellData['prices']));
+    }
+
+    protected function filterTable(Manager $I): void
+    {
+        $billPage = new IndexPage($I);
+
+        $billPage->filterBy(Dropdown::asTableFilter($I, 'Type'),
+            '-- ' . $this->sellData['type']);
+
+        $billPage->filterBy(Input::asTableFilter($I, 'Description'),
+            $this->sellData['descr']);
+    }
+
+    /**
      * @return array
      */
     protected function getPartData(): array
     {
         return [
-            'partno'        => 'E5620',
+            'partno'        => 'CHASSIS EPYC 7402P',
             'src_id'        => 'TEST-DS-01',
             'dst_id'        => 'TEST-DS-02',
             'serials'       => 'MG_TEST_PART' . uniqid(),
@@ -153,6 +206,20 @@ class PartsCreationCest
             'price'         => 200,
             'currency'      => 'usd',
             'company_id'    => 'Other'
+        ];
+    }
+
+    protected function getSellData(): array
+    {
+        return [
+            'sellInfo' => [
+                'contact_id'=> 'Test Manager',
+                'currency'  => 'eur',
+                'descr'     => 'test description ' . uniqid(),
+                'type'      => 'HW purchase',
+                'prices'    => [250, 300, 442],
+                'time'      => (new \DateTime())->format('Y-m-d H:i'),
+            ]
         ];
     }
 }
